feat: setup infrastructure and recovery wrapper for update system
